+++ test/sql/init.sql
@@ -3,5 +3,10 @@ CREATE EXTENSION ddlx;
 
 \dx ddlx
 
-select oid::regprocedure,obj_description(oid) from pg_proc p where proname like 'ddlx_%'
-order by obj_description(oid) is null, cast(oid::regprocedure as text) collate "C";
+with
+a as (
+ select oid::regprocedure,obj_description(oid) from pg_proc p where proname like 'ddlx_%'
+ order by obj_description(oid) is null, cast(oid::regprocedure as text) collate "C"
+)
+select row_number() over() as i,* from a
+;